chore: moved class ModelTelemetry and utilities to telemetry.py
